requirements.txt
Flask==2.3.3
pymongo==4.6.0
psycopg2-binary==2.9.10